+++ commands/README.md
@@ -35,9 +35,7 @@ const extension: JupyterFrontEndPlugin<void> = {
       label: 'Execute jlab-examples:command Command',
       caption: 'Execute jlab-examples:command Command',
       execute: (args: any) => {
-        console.log(
-          `jlab-examples:command has been called ${args['origin']}.`
-        );
+        console.log(`jlab-examples:command has been called ${args['origin']}.`);
         window.alert(
           `jlab-examples:command has been called ${args['origin']}.`
         );
@@ -51,6 +49,7 @@ const extension: JupyterFrontEndPlugin<void> = {
       );
     });
   }
+};
 ```
 
 The CommandRegistry is an attribute of the main JupyterLab application
@@ -71,12 +70,12 @@ with the unique command id and optionally the arguments.
 ```ts
 // src/index.ts#L31-L36
 
-// Call the command execution
-commands.execute(command, { origin: 'from init' }).catch(reason => {
-  console.error(
-    `An error occurred during the execution of jlab-examples:command.\n${reason}`
-  );
-});
+      console.error(
+        `An error occurred during the execution of jlab-examples:command.\n${reason}`
+      );
+    });
+  }
+};
 ```
 
 When running JupyterLab with this extension, the following message should
